- moved BCO matching code to fun4allraw. - use BCO matching class for RawDataEvaluation.
