+++ offline/packages/micromegas/MicromegasRawDataEvaluation.h
@@ -10,6 +10,7 @@
 #include "MicromegasMapping.h"
 
 #include <fun4all/SubsysReco.h>
+#include <fun4allraw/MicromegasBcoMatchingInformation.h>
 #include <phool/PHObject.h>
 
 #include <TTree.h>
@@ -63,10 +64,6 @@ class MicromegasRawDataEvaluation : public SubsysReco
 
   /// output file name for evaluation histograms
   void set_evaluation_outputfile(const std::string& outputfile) { m_evaluation_filename = outputfile; }
-
-  /// set gtm clock multiplier
-  static void set_gtm_clock_multiplier( double value )
-  { bco_matching_information_t::m_multiplier = value; }
 
   class Sample
   {
@@ -257,61 +254,11 @@ class MicromegasRawDataEvaluation : public SubsysReco
   //! main branch
   Container* m_container = nullptr;
 
-  //! store relevant information for bco matching between lvl1 and fee.
-  using m_bco_matching_pair_t = std::pair<unsigned int, uint64_t>;
-  class bco_matching_information_t
-  {
-    public:
-
-    //! verified flag
-    /**
-     * the flag is set to true as soon as a non-trivial match is found between FEE_BCO and GTM_BCO
-     * it shows that the chosen reference are correct
-     */
-    bool m_verified = false;
-
-    //! number of gtm bco rollover
-    unsigned int m_gtm_bco_ovf_count = 0;
-
-    //! first lvl1 bco (40 bits)
-    bool m_has_gtm_bco_first = false;
-    uint64_t m_gtm_bco_first = 0;
-
-    //! first fee bco (20 bits)
-    bool m_has_fee_bco_first = false;
-    unsigned int m_fee_bco_first = 0;
-
-    //! list of available bco
-    std::list<uint64_t> m_gtm_bco_list;
-
-    //! matching between fee bco and lvl1 bco
-    std::list<m_bco_matching_pair_t> m_bco_matching_list;
-
-    //! need to truncate bco matching list to some decent value
-    void truncate( unsigned int /* maxsize */ );
-
-    //! get predicted fee_bco from gtm_bco
-    unsigned int get_predicted_fee_bco( uint64_t ) const;
-
-    // this is the clock multiplier from lvl1 to fee clock
-    static double m_multiplier;
-
-    // define limit for matching two fee_bco
-    static constexpr unsigned int m_max_fee_bco_diff = 50;
-
-    // define limit for matching fee_bco to fee_bco_predicted
-    static constexpr unsigned int m_max_gtm_bco_diff = 100;
-
-    // needed to avoid memory leak. Assumes that we will not be assembling more than 50 events at the same time
-    static constexpr unsigned int m_max_matching_data_size = 50;
-
-  };
-
-  /// map bco_information_t to packet id
-  using bco_matching_information_map_t = std::map<unsigned int, bco_matching_information_t>;
+  //! map bco_information_t to packet id
+  using bco_matching_information_map_t = std::map<unsigned int, MicromegasBcoMatchingInformation>;
   bco_matching_information_map_t m_bco_matching_information_map;
 
-  /// map waveforms to bco
+  //! map waveforms to bco
   /** this is used to count how many waveforms are found for a given lvl1 bco */
   using bco_map_t = std::map<uint64_t, unsigned int>;
   bco_map_t m_bco_map;
